Add MSVC v15 (2017) project, tidy/comment a bit, remove a dependency
Removed build dependency Cinder block Cinder-OpenCV, since the new
project file links with the current OpenCV 3.3, but the Cinder-OpenCV
included with fellow build dependency Cinder 0.8.6 seems possibly to be
tied to a version 2 OpenCV if the binaries included with that Cinder
version are anything to go by. The application only makes a small use of
OpenCV and the Cinder block anyway; the minimal utilities used from the
Cinder block have simply been copied in.